+++ build_runner/lib/src/io/asset_path_provider.dart
@@ -13,11 +13,11 @@ abstract interface class AssetPathProvider {
   /// Set [hide] to get a path in the hidden "build cache" folder instead of the
   /// directory containing manually written source code.
   ///
-  /// Set [checkDeleteAllowed] to throw if the path is read only.
+  /// Set [checkWriteAllowed] to throw if the path is read only.
   String pathFor(
     AssetId id, {
     required bool hide,
-    bool checkDeleteAllowed = false,
+    bool checkWriteAllowed = false,
   });
 
   /// Returns [id] hidden in [buildCachePackage].